Add files via upload
I really wanted to show the beauty of the world around us using real pictures of places around the world that almost don't look real. However, as I was looking at these pictures, I started to think about why we don't see the world like it is portrayed on the screen. It made me think about how a photographer can change things like the angles, lighting, and focus to make us see things that we normally would not see, to give us a new perspective. I made the pictures really big so that they would be very pixelated and hard to see, and I added the black circle that follows the mouse to create blind spots. This is all to show how most of the time, we fail to see the real beauty around us, either because we have trouble seeing it or because of how our actions can sometimes hide it.